sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slsr-sdk"
version = "0.1.0"
description = "Self-Learning & Self-Replication SDK — governed autonomous agents with energy budgeting, state tracking (S=Q·M·T), and π·cos(√e) convergence gating."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Abacus.AI Autonomous Systems" }
]
keywords = ["autonomous-agents", "self-replication", "governance", "reinforcement-learning"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.21",
  "ruff>=0.4",
  "mypy>=1.8",
]

[project.urls]
Homepage = "https://example.com/slsr-sdk"
Design = "file:///home/ubuntu/sdk_design/SDK_DESIGN_DOC.md"

[tool.setuptools.packages.find]
include = ["slsr*"]
exclude = ["tests*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
